% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{as.character.neverpasteDocument}
\alias{as.character.neverpasteDocument}
\title{Convert a Document to character.}
\usage{
\method{as.character}{neverpasteDocument}(doc)
}
\description{
Convert a Document to character.
}
