% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmarks.R
\name{bookmarks}
\alias{bookmarks}
\title{Produce bookmarks}
\usage{
bookmarks(...)
}
\arguments{
\item{...}{Named arguments that you want to be bookmarked. If no name exists, it will be given a noname name.}
}
\value{
The bookmarked values pasted together.
}
\description{
Produce bookmarks
}
