% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document.R
\name{Document}
\alias{Document}
\title{Produce a new document to work with. Use the \code{add} function, and when finished, use \code{build}.}
\usage{
Document()
}
\value{
A neverpaste Document.
}
\description{
Produce a new document to work with. Use the \code{add} function, and when finished, use \code{build}.
}
