% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmarks.R
\name{setup_knitr_bookmark_hooks}
\alias{setup_knitr_bookmark_hooks}
\title{Hook into knitr to automatically bookmark output}
\usage{
setup_knitr_bookmark_hooks()
}
\description{
Setup the hooks for the use of of this package in combination with `knitr`. Every inline chunk or regular chunk will be bookmarked.
In the case of regular chunks, the label of the chunk will be used.
}
